fix structure more
